// src/app/services/member/member.interface.ts
export interface MemberInterface {
    id: number;
    first_name: string;
    last_name: string;
    latitude: string;
    longitude: string;
    firma: string;
    address: string;
    city: string;
    state: string;
    country: string;
    zip: string;
    phone: string;
    email: string;
    website: string;
    branche: string;
    lehrberuf: string;
}